+++ src/commands/a32nx/weightBalance.ts
@@ -10,10 +10,10 @@ export const weightBalance: CommandDefinition = {
         const weightBalanceEmbed = makeEmbed({
             title: 'FlyByWire A32NX | Fuel and Weight',
             description: makeLines([
-                'We have recently improved the A32NX flight model and finalized our weight and balance loading integration available on the development version. Please note:',
+                'We have improved the A32NX flight model and finalized our weight and balance loading integration.',
                 '',
+                '**Please note:**',
                 '- We have disabled the MSFS fuel and weights UI to prevent issues.',
-                '',
                 '- Use the EFB for fueling and the AOC W/B Page on the MCDU for payload.',
                 '',
                 'Please read our guide [here.](https://docs.flybywiresim.com/fbw-a32nx/feature-guides/loading-fuel-weight/)',