+++ src/commands/index.ts
@@ -118,6 +118,7 @@ import { latlongfix } from './general/latlongfix';
 import { headwind } from './general/headwind';
 import { wolframalpha } from './utils/wolframalpha';
 import { birthday } from './utils/birthday';
+import { recommendedSettings } from './a32nx/recommendedsettings';
 
 const commands: CommandDefinition[] = [
     ping,
@@ -238,6 +239,7 @@ const commands: CommandDefinition[] = [
     headwind,
     wolframalpha,
     birthday,
+    recommendedSettings,
 ];
 
 const commandsObject: { [k: string]: CommandDefinition } = {};